addPhenocamUptimes {ytBangerz}	R Documentation

Append Uptime Report with Phenocam Data

Description

Append Uptime Report with Phenocam Data

Usage

addPhenocamUptimes(yesterday, uptime.report.file, stat.save.dir)

Arguments

yesterday
The date to collect phenocam data for. Does not literally have to be yesterday

uptime.report.file
The path to an uptime report file to add phenocam data to.

stats.save.dir
The directory where stat files are saved

Details

This function appends in phenocam data for a date to an existing uptime report. If phenocam data exists in the report, it will be overwritten by data from a fresh query. Uptime data are generated by direct query of the UNH Phenocam API.

Value

The selected uptime report will have phenocam data added.

Author(s)

Robert 'n' Josh

[Package ytBangerz version 0.1.0 Index]