applyRanges {ytBangerz}	R Documentation

identify percentage of out of range data for a given stream(s)

Description

This function generates a status.df for in/out of range percentages, and is generic to all IS data products

Usage

applyRanges(site, report.df, dp.id, l0.matrix)

Arguments

site
character class defining site of interest, e.g., "CPER"

report.df
class dataframe, comprising reports of IS streams by site

dp.id
character vector of full NEON Data Product Identifiers, e.g., "NEON.D10.CPER.DP0.00004.001.00474.000.035.000"

l0.matrix
matrix of L0 data products

Value

a site-specific dataframe comprising in/out of range % for all IS sensors for a given date

Author(s)

Ronbert Lee Jeffershun IV rlee@battelleecology.org

[Package ytBangerz version 0.1.0 Index]