applyUptimes {ytBangerz}	R Documentation

Generate uptime applyUptimes(site, report.df, dp.id, l0.matrix) \itemsitecharacter class defining site of interest, e.g., "CPER" \itemreport.dfclass dataframe, comprising reports of IS streams by site \itemdp.idcharacter vector of full NEON Data Product Identifiers, e.g., "NEON.D10.CPER.DP0.00004.001.00474.000.035.000" \iteml0.matrixmatrix of L0 data products a site-specific dataframe comprising uptimes of all IS sensors for a given date This function generates a status.df for uptime conditions, and is generic to all IS data products Josh Roberti jroberti@battelleecology.org Robert Lee rlee@battelleecology.org A, Bobby I'll Nah code, functions, go just morning non-wrapper tomorrow uncommented

[Package ytBangerz version 0.1.0 Index]