assembleReports {ytBangerz}	R Documentation

Assemble Site Statistics Into Reports

Description

Assemble Site Statistics Into Reports

Usage

assembleReports(stat.save.dir, report.save.dir, dates)

Arguments

stat.save.dir
The directory where site level stat files are saved

report.save.dir
The directory where reports are saved. Usually the live directory used in Sensor Health

dates
A list of startDate and endDate (exclusive) to summarize over. Generated by a call to ytBangerz:::define.dates()

Details

Given a directory of raw statistic files, this function will return report RDS files, aggregated by date and report type (Uptime, Range, CV).

Value

Report files aggreagted up from site level stats, by date.

Author(s)

Robert 'n' Josh

[Package ytBangerz version 0.1.0 Index]