getMinMeanMax {ytBangerz}	R Documentation

Get Min/Mean/Max values

Description

Get Min/Mean/Max values

Usage

getMinMeanMax(site, l0.data, streams)

Arguments

site
The 4-Letter site code for a NEON site. Please, no OKSR.

l0.data
a data frame of 1) timestamps, 2) stream IDs, and 3) values

streams
A list of streams expected at the site

Details

From a set of L0 data, extract all min mean max values for all expected streams.

Value

A data frame of min mean max values, along with site, English data product names, term names, English locations, and the generic stream string.

Author(s)

Robert 'n' Josh

[Package ytBangerz version 0.1.0 Index]