getSiteStreams {ytBangerz}	R Documentation

Return the Sensor Streams at A Given Site

Description

Return the Sensor Streams at A Given Site

Usage

getSiteStreams(site, l0.matrix, exception.df = NULL, date = NULL)

Arguments

site
The 4-Letter site code for a NEON site. Please, no OKSR.

l0.matrix
The l0.matrix as a data frame.

exception.df
A data frame of excepted streams. Full 45-character streams and site column required

date
A YYYY-MM-DD character date. Input a value to return sensors installed at a site on a given date.

Value

Report files aggreagted up from site level stats, by date.

Author(s)

Robert 'n' Josh

[Package ytBangerz version 0.1.0 Index]