groupBySensor {ytBangerz}	R Documentation

Group output by sensor group

Description

Function to subset the domain reports down to a specific subsystem.

Usage

groupBySensor(domain, sensorGroup, domainReportText)

Arguments

domain
character class, "All" or code, e.g. "D02"

sensorGroup
Character class, one of "tis.met", "tis.soil", "sae", "ais.met", or "ais.aquSensors"

domainReportText
Nested list of domain report text (from generate_domain_reports.R, see send_alert_emails.r for example of how to handle)

Value

text reports of human readable sensor health split by sensor-type

Author(s)

Robert Lee rlee@battelleecology.org

[Package ytBangerz version 0.1.0 Index]