normalizeTo75 {ytBangerz}	R Documentation

Assemble Site Statistics Into Reports

Description

Assemble Site Statistics Into Reports

Usage

normalizeTo75(site, date, stat.save.dir)

Arguments

site
The 4-Letter site code for a NEON site. Please, no OKSR.

date
A single date, in the format "YYYY-MM-DD"

stat.save.dir
The directory where site level stat files are saved

Details

Somtimes, the site LC is very slow to stream the last 25 normalization only happens if all values <75 RH is excluded from consideration due to it's messiness

Value

Site uptime stats are normalized (as appropriate) and written into the site's statistic files.

@export

Author(s)

Robert 'n' Josh

[Package ytBangerz version 0.1.0 Index]