parseAlerts {ytBangerz}	R Documentation

Parse out site-specific alerts from overarching large list of alerts

Description

this function takes a date and the directory of alerts and generates a list of nested dataframes comprising site-specific sensor alerts

Usage

parseAlerts(date, alert.dir)

Arguments

date
POSIX input in the form of YYY-mm-dd. Used to search for alert files for given date

alert.dir
directory where the alerts are located; this is initiated from a script located upstream of this function

Value

a nested list of data frames comprising site-specific alerts for the provided date.

Author(s)

Robert Lee rlee@battelleecology.org Josh Roberti jroberti@battelleecology.org

[Package ytBangerz version 0.1.0 Index]