parseReports {ytBangerz}	R Documentation

Parse out site-specific reports from overarching large list of reports

Description

this function takes a date and two directories, one comprising domain reports and the other comprising site-specific alerts and generates domain reports

Usage

parseReports(date, report.dir, site.alert.df)

Arguments

date
POSIX input in the form of YYY-mm-dd. Used to search for alert files for given date

report.dir
directory where the domain reports are located; this is initiated from a script located upstream of this function

site.alert.dir
directory where the site-specific alerts are located; this is initiated from a script located upstream of this function

Value

a site-specific dataframe comprising reports of IS streams

Author(s)

Robert Lee rlee@battelleecology.org Josh Roberti jroberti@battelleecology.org

[Package ytBangerz version 0.1.0 Index]