prestoQuery {ytBangerz}	R Documentation

Extract a time window of a list of NEON measurement streams (L0) from ECS using Presto

Description

Modified from code by Cove Sturdevant. Function definition. Extract a window of NEON (L0) data from the NEON database using the NOTE: Your computer must be connected to the NEON intranet.

Usage

prestoQuery(
  idDp,
  timeBgn,
  timeEnd,
  Srvr = c("https://den-devpresto-1.ci.neoninternal.org:8443",
    "https://den-prodpresto-1.ci.neoninternal.org:8443")[2],
  Type = c("numc", "str")[1],
  CredPsto = NULL,
  PrcsSec = 3
)

Arguments

idDp
String containing full data product ID (e.g. "NEON.DOM.SITE.DP0.00004.001.00474.000.035.000")

timeBgn
A POSIX format date-time in UTC, denoting the beginning date-time

timeEnd
A POSIX format date-time in UTC, denoting the ending date-time. Non-inclusive.

Srvr
A character string - the NEON server to extract from, including port. Options are "https://den-devpresto-1.ci.neoninternal.org:8443", "https://den-prodpresto-1.ci.neoninternal.org:8443". Defaults to prod.

Type
A character string of either 'numc' or 'str' defining the data type of the requested data. Defaults to 'numc'

CredPsto
<- A character string with the username and password used to authenticate connection to the database. Must be in the format user;pwd, where user is the username, and pwd is the password. If NULL (default) the environment variable CRED_PSTO will be called and parsed based on the same format (CRED_PSTO='user;pwd').

PrcsSec
A single numeric (integer) value indicating the operational precision of the seconds field of time vectors. Defaults to 3 (millisecond-precision). Values higher than 6 cannot be guaranteed to produce desired results.

Value

A data frame of: time = the POSIXlt readout times (UTC) data = the requested data stream (same length as time)

Author(s)

Cove Sturtevant csturtevant@neoninc.org Steve Jacobs sjacobs@battelleecology.org Josh Roberti jroberti@battelleecology.org

References

License: Terms of use of the NEON FIU algorithm repository dated 2015-01-16.

See Also

wrap.extr.neon.dp

wrap.extr.neon.dp.pdr

def.extr.neon.dp.s3

Examples

# Using environment variable CRED_PSTO for authentication
idDp <- "NEON.D04.GUAN.DP0.00041.001.01728.002.507.000"
timeBgn <- base::as.POSIXct("2017-09-01T00:00Z",tz="GMT",format="%Y-%m-%dT%H:%MZ") # Begin date to grab.
timeEnd <- base::as.POSIXct("2017-09-02T00:00Z",tz="GMT",format="%Y-%m-%dT%H:%MZ") # End date to grab.
Srvr <- 'https://den-prodpresto-1.ci.neoninternal.org:8443'
Type <- 'numc'
som::prestoQuery(idDp=idDp,timeBgn=timeBgn,timeEnd=timeEnd,Srvr=Srvr,Type=Type)

# Using explicit credentials for authentication
idDp <- "NEON.D04.GUAN.DP0.00041.001.01728.002.507.000"
timeBgn <- base::as.POSIXct("2017-09-01T00:00Z",tz="GMT",format="%Y-%m-%dT%H:%MZ") # Begin date to grab.
timeEnd <- base::as.POSIXct("2017-09-02T00:00Z",tz="GMT",format="%Y-%m-%dT%H:%MZ") # End date to grab.
Srvr <- 'https://den-prodpresto-1.ci.neoninternal.org:8443'
Type <- 'numc'
CredPsto <- 'yourusername;yourpassword'
som::prestoQuery(idDp=idDp,timeBgn=timeBgn,timeEnd=timeEnd,Srvr=Srvr,Type=Type,CredPsto=CredPsto)

[Package ytBangerz version 0.1.0 Index]