productStatus {ytBangerz}	R Documentation

Generate Status Data Frame for Data Products

Description

Generate Status Data Frame for Data Products

Usage

taatStatus(site, l0.matrix, report.df)

precipStatus(site, l0.matrix, report.df, creds, meanValues)

phenocamStatus(site, l0.matrix, report.df)

relHumidStatus(site, l0.matrix, report.df)

mfcStatus(site, l0.matrix, report.df)

mfmStatus(site, l0.matrix, report.df)

inletPressureStatus(site, l0.matrix, report.df)

samplingPumpStatus(site, l0.matrix, report.df)

baroPreStatus(site, l0.matrix, report.df)

parStatus(site, l0.matrix, report.df)

qlParStatus(site, l0.matrix, report.df)

windStatus(site, l0.matrix, report.df)

atmoCo2IsoStatus(site, l0.matrix, report.df)

atmoH20IsoStatus(site, l0.matrix, report.df)

buoyPressStatus(site, l0.matrix, report.df)

dustSizeStatus(site, l0.matrix, report.df)

ecseConcentrationStatus(site, l0.matrix, report.df)

surfaceWaterElevStatus(site, l0.matrix, report.df)

irBioTempStatus(site, l0.matrix, report.df)

hutEnvStatus(site, l0.matrix, report.df)

parAtWaterStatus(site, l0.matrix, report.df)

nitrateInWaterStatus(site, l0.matrix, report.df)

swLWRadStatus(site, l0.matrix, report.df)

particulateMassStatus(site, l0.matrix, report.df)

wetDepStatus(site, l0.matrix, report.df)

buoyWindStatus(site, l0.matrix, report.df)

waterQualityStatus(site, l0.matrix, report.df)

tempChainStatus(site, l0.matrix, report.df)

surfaceWaterTempStatus(site, l0.matrix, report.df)

shfStatus(site, l0.matrix, report.df)

soilTempStatus(site, l0.matrix, report.df)

soilCo2Status(site, l0.matrix, report.df)

condTempElevStatus(site, l0.matrix, report.df)

primPyranometerStatus(site, l0.matrix, report.df)

dirDifStatus(site, l0.matrix, report.df)

Arguments

site
The 4-Letter site code for a NEON site. Please, no OKSR.

l0.matrix
The l0.matrix as a data frame.

report.df
A data frame listing all relevant report values for the site at hand. Generated by parseReports. A data frame of just the rows of a report related to the site of interest, with report type and stream IDs added per each column. For example, if CPER is the site at issue, report.df is a report containg only lines for CPER, with the type of report and stream added on. >>> VERY IMPORTANT: Rather than beat head against formatting, please use parseReports(), a function in domain_report_framework.R. It requires a DF of alerts at a site, as well as a date and directory where reports live. Example of implementation in generate_domain_reports.R

Details

These functions return machine-readable summaries of sensor conditions at a given site. Under the hood, each function makes a (status.df). This can be generated with makeStatusDF, like so: makeStatusDf(dp.id = "DP0.00001.001", site = "WREF", l0.matrix = l0.matrix, conditions = c("Anemometer tongs blocked", "Heater broken", "Tempertature senosr/PRT broken")) The output of the above line yields:

stream                      condition 000.010 000.020 000.030 000.040 000.050 000.060 000.070
 01306       Anemometer tongs blocked   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
 01307       Anemometer tongs blocked   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
 01308       Anemometer tongs blocked   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
 01309       Anemometer tongs blocked   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
 01310       Anemometer tongs blocked   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
 01306                  Heater broken   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
 01307                  Heater broken   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
 01308                  Heater broken   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
 01309                  Heater broken   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
 01310                  Heater broken   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
 01306 Tempertature senosr/PRT broken   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
 01307 Tempertature senosr/PRT broken   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
 01308 Tempertature senosr/PRT broken   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
 01309 Tempertature senosr/PRT broken   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
 01310 Tempertature senosr/PRT broken   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE

After generating this table in the function, the function will refer to a report file passed to the function to change logical cases for stream/condition settings at each location. The script used to generate the verbose alerts passes a subset report (
report.df) into each status function. report.df is a tall and flattened format of the sensor health reports. The column 'type' refers to the type of report information on the row (options are uptime, range, and cov), while the other columns are as shown in other reports. Only one column for site information is preserved. As an example of the input format of report.df, here is a report.df generated for JERC:

  type                          streams                                            dp.name  term.name             location  JERC
  range  DP0.00024.001.01320.000.010.000                Photosynthetically Active Radiation      inPAR         Meas. Level 1  60.1
  range  DP0.00024.001.01320.000.020.000                Photosynthetically Active Radiation      inPAR         Meas. Level 2  61.6
  range  DP0.00024.001.01320.000.030.000                Photosynthetically Active Radiation      inPAR         Meas. Level 3  88.9
  range  DP0.00024.001.01320.000.040.000                Photosynthetically Active Radiation      inPAR         Meas. Level 4  80.1
  range  DP0.00024.001.01320.000.060.000                Photosynthetically Active Radiation      inPAR             Tower Top  60.4
  range  DP0.00024.001.01321.000.060.000                Photosynthetically Active Radiation     outPAR             Tower Top  59.2
  uptime DP0.00066.001.01329.003.000.000 Photosynthetically active radiation (Quantum Line)    linePAR           Soil Plot 3  60.6
  cov    DP0.00040.001.01800.003.501.000                               Soil heat flux plate curVoltage  Soil Plot 3, depth 1   1.7

See the script function in .../L0-tracking/scripts/generate_smart_alerts.R for specifics on how report.df is generated for each site. Otherwise only the l0.matrix and site being tested should be passed into the fucntion. Once specific conditions have been addressed, uptime status can be added to the status DF by calling applyUptimes(), and rbinding the resulting data frame to the existing status DF. At the end of the fucntion, rows with all NAs should be removed, for readability. Note that each condition is matched with all streams, leading to nonsensical parings. For example, u and v vector component streams are paired with "Heater broken". If the logical conditions used to assign values are correctly mapped to condition and stream, these non-sense pairings will never be TRUE. Thus, removing all FALSE lines will remove these instances, and lead to a more sensible output. PRO TIP - use "som::def.neon.term(idDpMain = "DP0.00006.001")" to return terms in a DP.
Value

A data frame with columns stream, condition, and hor.ver locations. Rows should be specific conditions for each stream/combination pair.

Author(s)

Robert 'n' Josh

[Package ytBangerz version 0.1.0 Index]