translateStatusDF {ytBangerz}	R Documentation

Translate Status DF (TRUE/FALSE) dataframes into human readable interpretations for daily domain reports

Description

This function generates human readable intepretations of the information within the status DFs. Used for domain reports

Usage

translateStatusDF(status.df, stream.ids, report.df = report.df)

Arguments

status.df
class dataframe comprising sensor statuses for a given site-date

stream.ids
character vector of stream identifiers with known issues

report.df
class dataframe, comprising reports of IS streams by site

Value

text output describing issues, e.g., failing, for a given sensor and location at a site

Author(s)

Rev. Robert Rishard Lee Jr rlee@battelleecology.org

[Package ytBangerz version 0.1.0 Index]