uptimeFromBulk {ytBangerz}	R Documentation

Distill Uptime Information from Bulk Statistics

Description

This function returns uptime values for streams, and adjusts uptimes for non-1Hz streams

Usage

uptimeFromBulk(bulkStats, irregularStreams)

Arguments

site.stats
a data frame of site statistics, generated by getSiteStats_fast

non.1hz.dp
class dataframe, a list of streams and conversion factors for uptimes

Value

The bulk statsitics data frame with an uptime column

Author(s)

Ronbert Lee Jeffershun IV rlee@battelleecology.org

[Package ytBangerz version 0.1.0 Index]